+++ src/spring/azext_spring/log_stream/writer.py
@@ -2,15 +2,14 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
-import sys
 
 
-class DefaultWriter:
+class DefaultWriter:  # pylint: disable=too-few-public-methods
     def write(self, data, end='', file=None):
         print(data, end=end, file=file)
 
 
-class PrefixWriter(DefaultWriter):
+class PrefixWriter(DefaultWriter):  # pylint: disable=too-few-public-methods
     def __init__(self, prefix):
         self.prefix = prefix